adaptivedemux: add "select-bitrate" signal to control ABR
Add a signal "select-bitrate" that allows an application to take
control of the adaptive bitrate logic in GstAdaptiveDemux.

This allow applications using GStreamer to have control of the ABR
algorithm without having to modify the gstadaptivedemux.c source code
and allows unit tests to provide repeatable tests.

To provide some information to a signal handler, a new message type
GST_ADAPTIVE_DEMUX_REPRESENTATION_MESSAGE_NAME is created that is used
by demux elements to post information about the available bitrates.

    https://bugzilla.gnome.org/show_bug.cgi?id=758387