hlsdemux: tests: add test of ABR with live playlists
Add a unit test to check how hlsdemux handles changing playlists
of live streams.

At startup of a live HLS stream, gst_m3u8_client_update will select a
starting position 3 fragments from the end of the file. If hlsdemux
performs a bitrate switch, it should not re-enforce the rule
about starting 3 fragments from the end, as this would cause it to
download the wrong fragment.

For example, consider the case of two bitrates containing 6 fragments:

	 800  0  1  2  3  4  5
	1500  0  1  2  3  4  5

At startup, the 800KBps stream is selected and fragment 3 is selected as
the start position. After downloading 3 and 4, hlsdemux switches to the
1500KBps stream. At this point it must not enforce the "3 fragments from
the end" rule, as that would cause it to download fragment 3 again (at
the new bitrate).

    https://bugzilla.gnome.org/show_bug.cgi?id=758387