use macro to define names of request and response structs
Following the Don't Repeat Yourself principle, define macros
for the structures that contain the request and response headers,
so that the name is not repeated in multiple places in multiple files.

Use gst_structure_new to create the response headers and add the date.
There is no need to perform this in two steps.