dashdemux: tests: remove contentProtection event from test engine
DASH specific information should not be in the generic adaptive demux
test engine. However, there is no other convenient place to store
per-stream information. The compromise is to use a field in the
test case data, that is able to hold information about every stream.
This is used to store the count of content protection events.