adaptivedemux: tests: fix usage of gst_test_clock
Fixed several problems with using gst_test_clock in adaptivedemux tests:

1. The design must ensure that the clock must be available for every call
of gst_adaptive_demux_update_test_clock. It already does, so used an
assert to check for this.

2. gst_test_clock_process_next_clock_id must be called only when the clock
was updated. Otherwise the function will be called for an old value of the
clock and it will seem that it triggered before its time.

3. gst_test_clock_process_next_clock_id must always return an id. Used an
assert to check for this.

4. clock_update_id is set when pipeline is started, so it must be set
when it is removed at the end of the test. Used an assert to check for this

5. g_source_remove (priv->clock_update_id) should be called before resetting
the system clock and priv->engine.clock. It is obvious that it must be done
before resetting priv->engine.clock. It can be done after resetting the
system clock to NULL, but it will be much cleaner if it is done before that.