dashdemux: fixup of polling UTC timing
The commit 3c2d839cb9fa1bc1a86922707558174281125700
(adaptivedemux: use GstSystemClock to all real-time calculations)
introduced a regression in the clock drift compensation, that causes
the timing source to be polled every time the manifest is refreshed.
If the manifest has a very small minimumUpdatePeriod, this can cause
significant numbers of requests to the clock server.

The commit fixes this issue, and makes the code slightly easier to
follow by using GstClockTime for all of its calculations, rather than
using microseconds for the next_update.